+++ src/pages/additem.js
@@ -1,0 +1,11 @@
+import Header from '../components/Header';
+
+function Additem() {
+  return (
+    <>
+      <Header />
+    </>
+  );
+}
+
+export default Additem;
